[feat]storm开发 实时匹配关键字

+++ dlmws-storm/src/main/java/com/yipingjian/dlmws/storm/entity/Rules.java
@@ -7,5 +7,5 @@ public class Rules {
     private String keywords;
     private String type;
     private Integer threshold;
-    private Long interval;
+    private Integer interval;
 }
